+++ backend/src/test/java/com/NBE4_5_SukChanHoSu/BE/domain/user/service/UserMatchingServiceTest.java
@@ -1,0 +1,7 @@
+package com.NBE4_5_SukChanHoSu.BE.domain.user.service;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class UserMatchingServiceTest {
+
+}
